Rename and improve readability
